+++ micro-ui/web/micro-ui-internals/packages/modules/core/CHANGELOG.md
@@ -35,6 +35,13 @@
 - Login and SignUp Pages UI for Sandbox is updated
 - Sandbox Product Details Page Updated 
 
+## [1.8.42(Sandbox)] - [22-Sept-2025]
+- Sandbox ui style changes (Landing Page/Product Page/Details Page/OTP Carousel) / CCSD-833
+
+## [1.8.43]  [26-June-2025]
+- Login and SignUp Pages UI for Sandbox is updated
+- Sandbox Product Details Page Updated 
+
 ## [1.8.42] [18-Jun-2025]
 - Updated the loader logic in login page